Revert "MediaHTTPConnection: move connection states into an inner class"
This reverts commit 21cb690e974961d8274a5f650f7d0dd5f37a2a75 [formerly 621e7968adf0253d5e22406f02ccc8bcc0eda5ec].

Many of the fields that were moved are annotated @UnsupportedAppUsage,
so the CL would have had undesirable app compat impact. Further,
because investigation has revealed that lock contention *is* possible,
we need to always acquire the lock anyway so there is no longer a
benefit in keeping all of the mutable state in a single field that
can be atomically set to null.

Test: Treehugger
Change-Id: I202c5653cb086d99228491e161a159bad640105a

Former-commit-id: dc9f4b4d5d28fc68b1b5e4e8500bf67d4b11621d