Add logs for power key in interceptKeyBeforeQueueing
We have a bug that power key is not being called. The issue is somewhere
before "powerPress". For a similar bug, we have added logs into
"powerPress" previously, but in the new bugreports, we are seeing that
even powerPress is not getting called.

To bisect this further, add log to interceptKeyBeforeQueueing for the
KEYCODE_POWER.

Bug: 128933363
Test: adb logcat, then press power key after display is ON. observe the
logs

Change-Id: Ib68e37aa403af965ea650a21373d3ffc7370b38c

Former-commit-id: 24201e6ec62a1d8ead600a1ad75fe41c23267844