Add SD card support for AOT
Some low end devices do not have enough space in internal storage
and need a SD card for the AndroidOnTap to work. This CL makes AOT
to use SD card when it exists. Also it adds a "os.aot.path" system
property to overwrite the path for testing.

Bug: 128801963
Test: adb shell am start-activity \
-n com.android.dynsystem/com.android.dynsystem.VerificationActivity  \
-a android.os.image.action.START_INSTALL    \
--es KEY_SYSTEM_URL file:///storage/emulated/0/Download/system.raw.gz  \
--el KEY_SYSTEM_SIZE $(du -b system.raw)  \
--el KEY_USERDATA_SIZE 8589934592

w/wo the "os.aot.path" system property and logcat

Change-Id: I1b35b39b207a68236bc44f8b3c4b952c7dbd5a1e

Former-commit-id: dc0ebba5103cbe3ed6e6d465918532e89993f90b