Correctly samples just around the navigation bar handle
The sample rect is 10dp left and right to the handle while taking full
size of the current navigation bar inset. Also changed the threshold to
pass the median luminance of 0.5f while only applying the color change
if the luminance change is above 0.05f. Also shrunk the width of the
navigation bar handle.

Fixes: 128937537
Bug: 112934365
Test: manual
Change-Id: Iac76ac8c09b70b803de427960710933f4eaf2516

Former-commit-id: ec1d8adbdcdfb3a708b64d02a2f859c8ea0d53a8