Fix LocationOnScreenTests failure
The issue instroduced by CL [1], mis-understood
isSelfAnimating meaning that will also check if the transition is wait
for animating, but it will break the test that use no_animation theme,
that caused delayed will set as true in commitVisibity then
didn't callback onAppTransitionFinishedLocked -> onEnterAnimationComplete
for activity.

Revert to use isRealAnimating to check if need delay to unblock the test
fail.

[1]: I83504d578a17856623a53c04a7d3c54e5bcab5f4

Fix: 129359550
Test: atest LocationOnScreenTests

Change-Id: I913fff0afa6dad1aab83c1f7ad026e3d9e4d56e5

Former-commit-id: d14173e55ac695ce92a94e638b458b3669938b4d