Get rid of some ancient logic around layouting invisible windows
The logic didn't make sense to me, so I tracked it down to a very
old CL Ie8d49fd050442ebbdcf0b805087894e3a2fc4be9. The justification
on the CL and on the bug seem outdated: Whenever a window becomes
visible, we guarantee that we do a layout with it, so any inset or
frame changes will be dispatched to the app, so we don't need to
layout it when they are invisible.

Test: go/wm-smoke
Test: Presubmit + various apps with rotation
Bug: 129389526
Change-Id: I82b5a778aba355b25e2f7aa884b47d98de5b0fc5

Former-commit-id: 91d382a318d36342447d7ffa6c98b4e061a3bbe0