Fixes edge back working assuming boot up has back button
Can execute from back gesture is determined by overview proxy via
launcher however upon boot without booting into launcher, the default
value for back is hidden and anything that boots into a screen that is
not launcher will not be able to run back even if the back button would
have been there in 2/3 button mode. Set the default to visible alpha
until launcher appears and sets it to 0.

Test: boot with kiosk mode with gesture nav and use back
Fixes: 128971665
Change-Id: I19e684354aac9d05b91f4a50f587d75a1274bf90

Former-commit-id: 33b5e9885ab04247f5ad8d681dfe88dea44046d7