Check if product app is unbundled from property.
Some of devices launched with old version of android (and maybe not
treblized) can come with no /product/lib. However in current we are
checking if product app is unbundled by checking /product/lib from
search path, this can lead to wrong check of unbundled. To fix this, I
created new property (ro.product.apps.unbundled) to decide if product
app is unbundled.

Bug: 129011845
Test: m -j && tested from crosshatch
Change-Id: Id30595b3ca88ffe008ffe479406d59b840b19b44

Former-commit-id: 22557fe696fcb234636cf39da1552d568445c6ba