Notify PacakgeWatchdog when apps are ready
To observe packages with explicit health check enabled,
PackageWatchdog needs to bind to the ExplicitHealthCheckService
living in ExtServices. During boot, it needs to wait for the package
ready signal before attempting to bind to ExtServices.

Bug: 120598832
Test: Manually tested that after an APEX update, the network stack
does not pass the explicit health check until WiFi is connected
successfully. If Wi-Fi is never connected and the network stack
monitoring duration is exceeded, the update is rolled back.

Change-Id: I9ebeced010381c958aa27a465f508c54502c6b47

Former-commit-id: 7b56e5607c9ed778379f88667ac76a9bf930adbd