Keep IME target till window is closing and not removed
Issue:
When current IME target window has IME open and user navigates
to next window, IME target is set to next window
(unless it's isActivityTypeHome).
This causes IME to be reparented to the next window. This next window may
not have an editor and IME will be dismissed.
This causes flicker: IME animating out as next window is animating-in.

The fix:
If we hold on to IME target till last IME target is closing and
not removed, the flicker can be prevented.

The restriction for isActivityTypeHome() was originally added for bug 78214125
which made sense prior to reparenting.

Fix: 129270541
Test: 1. Manually tested steps in bug and flickering was not observed.
      2. Verified with next window having IME up (stateAlwaysVisible).
         No noticeable change.
      3. also ran go/wm-smoke-auto

Change-Id: I016c1e4549cf03c09cf1a30a91bd55fdb2587075

Former-commit-id: 2d0aca48c049f7c3069ba64475ad25255f70fbbc