Notify callbacks to update in case wallpaper is waiting for unlock.
If the wallpaper isn't direct-boot aware, the fallback wallpaper will be
bound first until we unlock from lock screen and then the desired one
will be bound.

However, the attributes will not be updated after the desired wallpaper
is bound, this might cause some bugs so we also notify callbacks in this
case.

Bug: 127810511
Test: Install 'Friendly Bugs Free L.Wallpaper' from play store.
Test: Use 'Friendly Bugs Free' as wallpaper.
Test: Press power key to enter AOD, observe the behavior.
Test: Restart device, unlock, press power key, observe behavior.
Change-Id: I7b8f44a0bcf452dcd757786dd12efa9eef22bae1

Former-commit-id: 96dbbb9d2e5c543bcf1e5b3a45229e479dd7f8f2