Allow background activity starts to be controlled by DeviceConfig.
If the global setting has been explicitly set to 1 or 0 (e.g. via
developer settings) then that takes precedence - so we don't interfere
with what the user has set. Otherwise the default value from
DeviceConfig is used.

Also migrate the associated (not yet used) package whitelist to
DeviceConfig, so we can set both at the same time. The whitelist is
ignored if the user has explicitly enabled or disabled background
starts.

Bug: 129533810
Test: atest WmTests:ActivityStarterTests

Change-Id: I2856edb5cb8c99a8cfef4712732d9dc9c5d7cdb7

Former-commit-id: fb227077b8652b817a094d37a5215ad7466ee603