Adding subscription group owner.
Define owner of a subscription group. Such that when subscriptions are
being added into or removed from a group, as long as callingPackage is
the owner, it can skip carrier permission check on the subscriptions in
the group.

This is necessary when a subscription is being added into a
group that contains inactive subscription, which mean its carrier
privilege can't be verified.

Bug: 131854492
Test: cts, unittest, TelephonyManagerTestApp
Change-Id: If2f5b9c8d71469d9e60f1ca48cab36938cd72c48

Former-commit-id: f81aae09817318b3a4ca9f61e85a3e7747df7709