Enforce new bubble policy in NoMan
If the app is foreground & the notification wants & is allowed to bubble
then we bubble.

Otherwise, the notification only gets flag bubble if it:
- is messaging style OR
- phone call with foreground service with person

If the app is foreground & something that typically wouldn't be allowed
to bubble gets to bubble, as long as that bubble is around, subsequent
updates to it will also be flagged as a bubble. Once the bubble / notif
is removed, subsequent notifs will not bubble.

Adds a tests for these various requirements.

Test: atest NotificationManagerServiceTest
Bug: 129147774
Change-Id: I27d28566af4c29ec59430939f82ac2fcedeb75fe

Former-commit-id: be7979616c74493d704bd57f8bcf678cd62c068f