Focus content or button on AlertController
Requests focus for the content view, or if the content does not accept
focus, a button on an alert dialog if not in touch mode.

Before, the ScrollView containing the buttons received focus, and the
user had to navigate to a button before being able to click it.

Tries to focus the first available button in the order of positive
button, negative button, neutral button.

Test: Manual test on Android TV:
       - Suspend YouTube package using `adb shell pm suspend com.google.android.youtube.tv`
       - Try starting the YouTube app, observe focus on alert dialog
       - Also tested with AlertDialogs with custom views, list view
Bug: 129467676
Change-Id: I69f39d0ade5a8b275f865ef1a76984407413c282

Former-commit-id: b96ca99b49e7fc7cbdbe7e98b376959173cc51b8