Ensure that THM sets up even when system dies
Sometimes, something crashes unrelated to Test Harness Mode on the first
boot. When this happens, if Test Harness Mode is partially set up, it
leaves the device in a state without ADB access. This change keeps the
persistent data around until after the device is set up entirely, so
that if the system crashes, it will still apply everything correctly.

Test: make, then run cmd testharness enable 100x to check for robustness
Change-Id: I4daa112f85e3234ab18a3aa1dae883cff61497df
Bug: 129769105

Former-commit-id: de0880a5068272bec6e870859f5c8701abb306fe