Ensure consistency of screenHeightDp for non-resizable activity
Originally the height of status bar (or other stable insets) is not
excluded from screenHeightDp for size compatibility mode activity
because the previous display where the activity resides may be removed
or changed then there is no available and consistent display bounds to
calculate the screen configuration.

Now we precompute and store the display insets in each rotation, so the
activity always has the original display insets to compute consistent
screen configuration.

Test: atest ActivityRecordTests
Test: atest TaskRecordTests#testComputeConfigResourceOverrides
Bug: 112288258

Change-Id: I5c769f8080fc5ae4eba966d2fa42cb650254074d

Former-commit-id: 61987bcaf58e5fca1b908b827784b8235c94ffaa