Temporary hack until we have dynamic remounting.
We're stuck in a place where we've half-implemented our storage
mounting and remounting story, so the mount mode is locked in for
the lifetime of the process.

The safest way to get this working for the next beta cut is to
always assume the app has the permissions they request, and we'll
follow up with a STOPSHIP to get a real fix in place.

Bug: 129744277
Test: atest --test-mapping packages/providers/MediaProvider
Change-Id: If152c52d3630112547f4d3e7a07449da452550fd

Former-commit-id: cf1fd5195f10e460f3efa2dd4b77ce3407c11443