Restore gnss@1.1 HAL capability bits removed in gnss@2.0 (framework)
In the IGnssCallback.hal@2.0 introduced in Android Q, the
capability bits in IGnssCallback.hal@1.1 that represent sub-HAL
interfaces have been removed as they are derivable from the
existing getExtensionXXX() family of methods in the IGnss.hal
interface.

These need to be restored back as the synchronous nature of the
getExtensionXXX() methods called by the framework has an impact on
partner implementations that need to communicate with the modem to
get the capabilities.

Additionally, the capability bit MEASUREMENT_CORRECTIONS needs to be
added for the new optional measurement_corrections@1.0 sub-HAL
introduced in gnss@2.0.

Fixes: 129870126
Test: Verified through cuttlefish default implementation and VTS tests.
Change-Id: I30c34361882e0722240feb3deba7d80136cd4bb3

Former-commit-id: 62c42dc1757ad6c1cc6597610622276eaa752409