Fix mixed up variable in triggering logic
When this code got refactored the check for severe warnings
accidentally had the time remaining being compared against the percentage
value of the severe warning threshold. Unfortunately this meant that
it was impossible to mark the severe level warning as shown and could
result in multiple triggerings. This change makes it so everything
is compared with the appropriate type of value again.

Test: PowerUI Tests pass
Bug: 129730217
Change-Id: Ic8bdabf1c70e49d117210bb80f1ca0a0d7684617

Former-commit-id: 290496e94e35d1bc804aad4ef7f7d04ee5c86eea