Add CallLog provider loggging to check for app-ops denial.
If an app-ops denial occurs while inserting into the call log in Telecom
its not entirely obvious why call log entries are not being added.
Adding a log message to make it more clear when call log insertion is
silently failing so that this can be root-caused easier in the future.

Test: Compile, manual test to check for log messages
Bug: 130045570
Change-Id: I85d834990d02939cf8530145c4d27d0939d0f8f7

Former-commit-id: 8e99c7b6e20b95fa442557e8d7cbf69a345d2d15