Doing sublisting after deduplication
For message like "a@anrdoid.com b@android.com", we wanna show no
suggestion on them because it will be ambiguous (Two chips with
label "Send Email"). And this is achieve by deduplication.

Previously, we did sublisting before duplication, and thus
we failed to remove some duplications. Let's do it the another way round.

BUG: 130344529

Test: atest
Change-Id: Idb4ea439daca710ce6247611f97c657e6a1c1a1b

Former-commit-id: aa496d04b479754a27d2866b006e2cd6b7884718