Remove the bubble if the notification is no longer FLAG_BUBBLE
Previously if a notification that was a bubble got updated such that it
would fail our bubble criteria, we wouldn't actually remove the bubble we
would just stop updating that bubble. This CL fixes it so that we'll remove
the bubble in that case.

This CL also factors all of the 'shouldBubble' logic into Notification
InterruptionStateProvider.

Fixes: 128459529
Test: CTSVerifier tests in other CL; atest BubbleControllerTest
Change-Id: Ia2b53a25a28d53bd2bbaebe4fe89f50d8d46a9fb

Former-commit-id: aa8fef22b728541b2413435adbe69164a647c4d8