Make lifecycle events for bubble activity views report normally
There are some extra onPause / onResume events in ActivityView due to
SysUI calling startActivity in onActivityViewReady which happens a bit
before the display being turned on.

This CL posts the startActivity call which is enough to avoid the extra
lifecycle events.

This CL also adds new method on ActivityView to startActivity with your own
ActivityOptions (because posting causes activity transition to occur...)

Test: manual - have Bubbles test APK
             - create a new bubble
             - adb logcat | grep "BubbleActivity"
             - observe the logging to make sure lifecycle events behave
               normally
Bug: 130363466
Change-Id: Ia44d6033e5cff625222006632b7bdc4dc1e59e81

Former-commit-id: 60101c9f298e69ee2b5ee5f67943515aae79b187