Fix issue in attached layouting
When we forced the parent to draw behind the navigation bar,
overscan/display are extended towards the bottom of the screen,
which will cause attached windows to layout behind the navigation
bar.

To fix this, we intersect these frames with DF.RestrictedOverscan
in case we forced the window to draw behind to keep compatibility.

Test: AccessibilityWindowQueryTest
Fixes: 130231322
Change-Id: I66f5ad4fab1f173e5b61e7d61d1e0bb0dafe66e5

Former-commit-id: 36dcd2e34be3c1a8b48270a06c5e69162a92a2e2