Add -i com.android.shell when --enable-rollback is specified.
com.android.shell has the TEST_MANAGE_ROLLBACKS permission needed to enable
rollback for non-module packages, which is likely what the user wants when
enabling rollback through the shell command. Set the installer to
com.android.shell if no installer has been provided so that the user doesn't
have to remember to set it themselves.

Bug: 129835930
Test: adb install RollbackTestAppAv1.apk
      adb install --enable-rollback RollbackTestAppAv2.apk
      adb shell dumpsys rollback, confirming a rollback is available for the
      test app.

Change-Id: I9f42cf0e1dab2efffb6badd6b8ea477887176bca

Former-commit-id: 88184a6be9461187fe989482d6b6c8247140eb94