Fix excessive entrance animation duration + delay
We were previously staggering animations by using the child's
notGoneIndex, but that can lead to strange results in the case of groups
that have many (~40) children. The children count towards the
notGoneIndex, causing unexpected discontinuities between notifs that are
visually right next to each other. Instead we just use a simple stagger
counter that gets reset on every pass and incremented for each new
entrance animation.

Fixes: 130259081
Test: Manual
Change-Id: I3c8519a8f9b7fcb661b2c4c4f8e85db0a717bc01

Former-commit-id: 4592f66058518419f53db421ae40e317696532d5