Restore window to undefined instead of freeform windowing mode
On freeform display, this will act in the exact same way as before. When
click the restore button to switch windowing mode, it will get into
freeform windowing mode.

This patch makes a difference when the display runtime changes from
freeform to fullscreen. If the window restored to undefined, when the
runtime change happens, it can properly match the parent's windowing
mode. Otherwise, we will see a freeform window on fullscreen display
without any explicit user action.

Bug: 128853293
Test: manual
Test: go/wm-smoke
Change-Id: I12480dedbdea3da65f7fa925eed0fa5664c76f5b

Former-commit-id: e9c1c319c70ab30aa53c029b6d54342230df3725