Move setProcessGroup outside oom adjustment
To reduce the time spent in applyOomAdjLocked. In the case that
when updating oom adjustment and apps call APIs which need to
acquire AMS lock, the duration of lock contention could be less.

Also remove the unnecessary clearCallingIdentity (commit 9438de2)
because the execution won't be on binder thread. Actually the
implementation of setProcessGroup doesn't look like there are any
binder related checks.

Bug: 123043091
Bug: 130295084
Test: AppLaunchTest

Change-Id: I4b593b4eeb647ce93182a9b776a664ed4414ce1d

Former-commit-id: 481d1e870e76e0ec7ce01ed1e09d0d1203465527