Fix colors in QS Header
Fixes icon and text colors in QS Header to be darker in Light mode.
Also, makes sure that all colors in the icons match.

Does this by implementing the following:
* A new theme that overrides darkIconTheme in Theme.SystemUI by one with
darker colors.
* QuickStatusBarHeader applies that theme and passes is onto its
children, in particular StatusIconContainer/StatusBarMobileView
and BatteryMeterView.
* StatusBarMobileView and BatteryMeterView retrieve the colors from the
theme and apply them accordingly.
* Also applies to QSCarrierGroup
* Additionally, the single color is used for the other icons.

Also invalidates SignalDrawable after setColors is called if at least
one of the colors changed. This is a real issue.

For more info:
* Text and icons set to 70% black on light mode, white in dark mode.
* Dual tone set for background overlaid with fill match single tone.

Test: visual
Fixes: 124466915

Change-Id: Ia6cb28fca90a07c8936726334502e7accf534df6

Former-commit-id: f86df994fa7b4f649bc6be551b6897fe82108e71