Do not populate SLOT_KEY with wrong slot info
The slotId is not necessarily the phoneId, so do not broadcast as if it
is. SLOT_KEY is a hidden extra so this should not impact apps in general
(if they need the old extra they can still use the hidden PHONE_KEY).

For select broadcasts like SIM_CARD_STATE_CHANGED, we will manually
populate the slot key. This is a temporary solution to b/130664115 and
we should eventually review and consider making the SLOT_KEY public for
some broadcasts.

Bug: 130664115
Test: manual
Change-Id: I84bd4c6bc4277e32cee9916dd18dd80948a808a0
(cherry picked from commit d64fc67a94e719c6e6c0a2443077041328fbd7e7 [formerly 21c3aeb49de064b011a452143f99b623acb94786])

Former-commit-id: 656f6f8b664e1dc7ca659fe60565f28979ec3e65