Do not populate SLOT_KEY with wrong slot info
The slotId is not necessarily the phoneId, so do not broadcast as if it
is. SLOT_KEY is a hidden extra so this should not impact apps in general
(if they need the old extra they can still use the hidden PHONE_KEY).

For select broadcasts like SIM_CARD_STATE_CHANGED, we will manually
populate the slot key. This is a temporary solution to b/130664115 and
we should eventually review and consider making the SLOT_KEY public for
some broadcasts.

Bug: 130664115
Test: manual
Change-Id: I84bd4c6bc4277e32cee9916dd18dd80948a808a0
Merged-In: I84bd4c6bc4277e32cee9916dd18dd80948a808a0

Former-commit-id: dd50042fb8578350cf24aed0ae8b66720e1c14a5