Fix mLaunchTaskBehind is not restored when canceling recents
If a 3rd party launcher is set as default home, when pressing
home button to go back from app, the launch-behind of recents
activity will be set. And then 3rd party launcher will move
to top that triggers stack order changed to cancel the recents.
That includes to restore launch-behind but the current top is
not the original one that was set. So there are 2 activities
show at the same time.

With storing the original target activity, we are able to
restore the launch-behind state correctly.

Bug: 131589476
Test: atest RecentsAnimationTest#testSetLaunchTaskBehindOfTargetActivity
Change-Id: Ib923492bb6263482fa22bd031a716b48ded20f91

Former-commit-id: 43233b72acba8ad4554501cd4f5a7860d9a009fd