Populate scrollX and scrollY on AccessibilityEvents.
This wasn't happening for RecyclerView, and that resulted in lost focus after scroll animations.
This is cause Talkback relies on changes in getScrollX/Y to determine if it should do anything about a TYPE_VIEW_SCROLLED AccessibilityEvent.
This squashes a big subset of "losted focus" bugs.
This is most painfully felt when you fling a RecyclerView and in ViewPager page transition, both resulting in slightly longer animations.
Also, change the default value of m[Max]Scroll[x/y] to reflect the above.

Change-Id: Ibe66260fbfc61c98ca88e1b2d9552ed116e44c15
Fix: 125385883
Test: CTSAccessibility*, Tried a few RecyclerViews on device, and ViewPager2 sample app.

Former-commit-id: 002e3cf73af21c82a981bc592f693f7525a7c4f6