Only start WebView zygote if multiprocess enabled.
The fix for b/130305037 inadvertantly changed the behaviour of the
webview zygote for devices which have multiprocess WebView disabled: the
zygote was being started at boot time even if it would not be needed.

Only call the function to start the zygote if multiprocess is actually
enabled, to fix this.

Fixes: 131267737
Test: adb shell cmd webviewupdate disable-multiprocess; adb reboot
Change-Id: Iccda48b75d75259821dba6b002339b2d6c8e3394

Former-commit-id: 1bc38eda65a9270d555967f3e00575817878e6e8