Rollback SHELL UID check for CarrierConfigLoader
When a test calls adoptShellPermissionIdentity before
calling a command, it doesn't adopt the shell UID, only
the permissions. So, we can not use
Binder#getCallingIdentity to determine it the API is
being called as part of a test. For now, we must only
use MODIFY_PHONE_STATE until a better pattern is
established.

Test: Manual
Bug: 131185660
Change-Id: I59d32a2e36950b2485e1a46a25e3724928676219

Former-commit-id: cf297467e062734ba19c02bdd879d6dd14769677