Remove the cache size from internal size.
Because the system can automatically clear away cache data, we remove
usage of it from the list of app storage usage in Settings. This should
encourage developers to more aggressively use the unattributed cache,
rather than the attributed regular data folder.

This affects the following fields on the AppEntry: internalSize & str
and size & sizeStr (which are derived from the internal size).

Currently, ApplicationsState.AppEntry.internalSize, externalSize, and
size are only used in the Settings apps so this shouldn't have broader
implications.

Test: Manual verification of Other Apps page
Fixes: 131152171
Change-Id: I3eb2e28701490a543c633e9c95a280059bb99c21

Former-commit-id: 02917dd5f716a9e95fb60a23580c12b0b75de573