Do not execute app transition too early
We execute an app transition if something was moved to the front,
but that activity was already resumed because when moving a task
to the front, we rely on the activityResumed callback to update
visibility and execute the transition usually, which doesn't
happen in that case.

However, we should only do that for the top task/stack as
otherwise we may execute an app transition too early in case an
activity behind the top one is still visible and resumed, like it
can happen when Launcher finishes the recents animation and
quickly launches the next one.

Test: Quick switch 100 times, no flashing
Bug: 129729087
Change-Id: Id76643e01c8e1c4b3ffd0d5a2fa38a35336e1d9e

Former-commit-id: a5cf6804fc4b78265e84d69e139f9b9f893a89c0