Fix display-level windowingMode change
This includes the following fixes:

It updates the ATM hierarchy and executes accumulated transitions
when display windowingMode changes. This was being skipped because
the mode change was being set on DC before reconfigureDisplayLocked.

Don't freeze the display when only window-effecting configs change

Don't cancel change transition on display change if the window
hasn't moved to a different display

Cancel change transition when app loses visibility.

Bug: 130755818
Test: Added test to AppChangeTransitionTests, manual test on chromeos
      for display windowing-mode change.
Change-Id: Ic1a0427f3e5ba940dc4c12765395e8ca050b0baf

Former-commit-id: b1e75f7b2a5444bccae00ec803a6bad81789ba19