Make Native Watchdog trigger RollbackManager rollback
Make RollbackPackageHealthObserver poll the status of
ro.init.updatable_crashing 10 times with 30-second interval.
This happens after the reboot.
If the property is set, a RollbackManager level rollback is triggered
for the rollback including Module Metadata Package.

Also post RollbackPackageHealthObserver#onBootCompleted to the handler
thread.

Note: the plan is to make the polling time and number of polls
configurable via DeviceConfig.

Test: planted logs and made sure the polling thread is active every 30
seconds for 10 times.
Test: * install a staged mainline module update with rollback enabled
      * reboot
      * kill zygote process 4 times
      * verify that rollback has been committed

Bug: 128501975
Change-Id: Id5fb37dd853a6e83e0663edb2857e5a24a597d1a

Former-commit-id: c5805fb97341de58c1c533c7b089cbef9de7103b