Fix getActiveSource api in HdmiControlService
The previous logic does not take an unregistered Active Source as a
valid one. Change it to be able to return with an HdmiDeviceInfo with
unregistered logical address or when the device is not updated in the
local remote device list.

Test: manual
Bug: 131618184
Change-Id: I64c9cbcdb812c16b42daccfa8da87d502b7f733f
(cherry picked from commit bcde21b0c6be81432930fe53840af218af9748c8)

Former-commit-id: 8a9eae71f33e640afe44473e9db20a2294584a62