Remove burn in offset on lock screen.
Due to the burn in offset applied to the analog clock, it can overlap
with "At a Glance" and the lock screen message. This change helps a
little bit by removing the offset on the lock screen, which hides
more of the clock face under the notifications.

Note also that the burn in offset is removed from the default clock
face on the lock screen. So, this change is consistent with that
behavior.

Fixes: 131122783
Test: visual -- analog clock moves to screen center on lock screen.
Change-Id: Ie57bb0b5a41215d4f1b478aeb691e07eac125614

Former-commit-id: e9aa87b26dbf99ed0abff8ae0199959449a57b53