Mark smart replies as sent on notification update/removal.
To keep notifications for which the user has sent replies we have some
logic using 'lifetime-extenders'. The SmartReply lifetime extender
ensures that we keep notifications around at least while smart replies
are in the process of being sent.
However, before this CL we never marked smart replies as sent - after
sending them we would mark them as 'sending' but not 'sent'. With this
CL we mark smart replies as 'sent' on notification update and removal to
allow notifications with smart replies to be removed.

Test note: I'm not adding any unit tests in this CL - the existing tests
only use mocks (not fakes), and there's thus no easy way to ensure we
call stopSending() from one of the NotificationEntryListener callbacks.

Bug: 122140949
Test: manual
1. Receive message
2. Tap on smart reply -> smart reply sent
3. receive new message (so 'mark as read' button appears)
4. tap 'mark as read'
5. ensure notification removed.
Test: manual
1. Receive message
2. Respond using reply button (remote-input) -> reply sent
3. receive new message (so 'mark as read' button appears)
4. tap 'mark as read'
5. ensure notification removed.
Change-Id: I2f0edba65f81d5b6fb577e9f83374c404a99e3fb

Former-commit-id: 36fc266f402367aa443b42a6291dcbde7455f6f5