Prevent estimates from updating while callbacks are happening
It seems that for some people the estimate was being set to null
as part of the process to update the estimate as they were trying
to access it. This CL makes it so that we don't allow callbacks
to start if we are currently updating which will ensure that this
doesn't happen again.

Test: repro'd issue without change, issue gone with change
Bug: 131607084
Change-Id: I938d05a3977bb4bba2c5d28e800d94120b8425d2
(cherry picked from commit 402ed31631c0b31c9d8d97716812587588cfae7c [formerly af36fb3c1c5f100e8c53d77aa15c7253330bfb9e])

Former-commit-id: 373400a5d065df511c71375ef3335842ca54db49