Properly handle activity start tokens for unordered broadcasts
Unordered broadcasts are (sometimes) fire-and-forget, with no final call
back to the system process indicating completion.  This means that we
were leaking the tokens associated with a process's capability to start
activities while in a background state.  This is now fixed: we ensure
that unordered broadcasts also have a capability-revoke operation
regardless of the behavior of the app process.

Along the way, we fix a latent potential leak that arose in the case of
delivery failure.

Fixes: 131309007
Test: sample app found at that bug
Test: atest WmTests:ActivityStarterTests
Change-Id: I9a2d7aba291cd2e52dd1ee7e9560a16ba5e42be5

Former-commit-id: 6ac3980275817a32847c95d16b5c3e484e253d84