Fixed that clicking on the remoteinput didn't work from AOD
Because we are freeing the AOD view, the view was not attached
to the window anymore and therefore we would not trigger the
remote input. Additionally if the expanded view was already showing
on AOD, we also wouldn't wait on it to become visible, also leading
to an issue where the user couldn't reply.
This also fixes an issue where the notificaiton would be invisible
if it was replied to in a group.

Fixes: 118365629
Fixes: 110072505
Change-Id: Ied33ed19eacc115006f5457f427acd4fefe0494b

Former-commit-id: f7d88932e2f699b99c1997a197c848d8e527f435