Prevent dark mode backup from applying in before setup
This CL makes it so that UIModeManager will not allow dark mode
to be changed prior to initial setup wizard being completed.
Additionally, it makes it so that when UIModeManager starts it
will try to start in light mode prior to setup wizard completing
regardless of what the backed up value is. We also register for
an observer if when we start UIModeManager we notice that setup
wizard has not run and defer updating dark mode to the true value
until we see that setup wizard completes.

Test: Manual
Bug: 119666724
Change-Id: Ie9cf36c8483d9cdd9e7aadd3aae4cc0ef15e8d45

Former-commit-id: c500b2741e31ce52356ff66e3b4943a98ae3ff8c