Fix transition animation of ResolverActivity for home
When touching navigation bar, it may trigger startRecentsActivity
to make the recents (home) activity visible to prepare recents
animation. If the final operation doesn't need the animation, the
animation will be canceled. When the condition is in the middle
of starting activity, the visibility of recents activity will be
updated after setting the transition of next activity. That leads
to the recents activity is added to closing apps so the transition
doesn't choose the animation style for translucent activity.

This change makes the recents activity invisible immediately if
it is really invisible to user, so its visibility change won't
affect the later launch transition.

Bug: 131589476
Test: atest RecentsAnimationTest#testRecentsActivityVisiblility

Change-Id: Ib8b59b9fc8dc40ea17d827a48f2b0cf2c1a1b3dd

Former-commit-id: 0e59172893e6d49b29f5efd4ce46c65fb2804b58