Restrict rotation when global actions panel is enabled
Once the orientation changes to portrait, the panel will be displayed,
and the orientation will be locked to portrait.

If, when global actions is launched and a panel is available, the user
has their orientation locked to non-portrait, then their orientation
is unlocked, so that they may rotate to portrait in order to see the
panel.

In all cases, the user's orientation setting is restored once global
actions is dismissed.

Bug: 129343749

Test: unlock orientation, launch global actions in portrait, notice
      that orientation is locked, dismiss global actions, notice that
      orientation is unlocked

Test: unlock orientation, launch global actions in non-portrait,
      rotate to portrait, notice that orientation is now locked,
      dismiss global actions, notice that orientation is unlocked

Test: lock phone orientation to non-portrait, launch global actions,
      rotate to portrait, notice that orientation is locked, dismiss
      global actions, notice that orientation has reset and is locked

Change-Id: I3972929aede075de69ea5333dcaf445e6ace59b8

Former-commit-id: 48f75dbbb6b3e4ad16ef88cf54544c15cbb0178b