Add BluetoothControllerImpl#onACLConnectionStateChanged
In some cases, a BT device can change its ACL connection state without
going through proper disconnection and then BluetoothControllerImpl
would not be updated without this. Listening to this changes makes
BluetoothControllerImpl aware of more state changes in BT state.

In the BR, there is an ACL_DISCONNECTED right before the screenshots
time, that does not have a CONNECTION_STATE_CHANGED so this seems to be
the issue.

Also added DEBUG logs for the different state changes

Test: atest BluetoothControllerImpl
Bug: 130058055

Change-Id: I40449dba4fb30673fc21ce6b9955d66261776ea2

Former-commit-id: 7f27b8ee653cc124ff4a931cc3d1f3c0a43ea3cf