Adds a PhoneStateMonitor to the AssistManager
This allows the AssistManager to determine the state of the phone when
the Assistant is invoked. This will be used for logging and allowing the
Assistant to alter its behavior based on the invokation state.

Note that SysUI is currently unable to differentiate between Launcher
states, so all Launcher states will be reported as PHONE_STATE_HOME.

Test: Tested locally
BUG:128982146
Change-Id: Iee40e42830f5a07845240c98bec5b87bd7c08eb7

Former-commit-id: a2c88760a7048d89f07e602dfbb5e15011d72cc5