Fix TileLayout#onMeasure
The width value for setMeasuredDimension was wrong, as it was adding the
padding, which was already included as this is the width we get from the
parent. This caused each TilePage in PagedTileLayout to be 8dp wider
than the container and then they overlapped at the edges, appearing in
the next page.

Test: visual using showLayoutBounds
Test: manual using Talkback, previous page not focused
Fixes: 132048870
Change-Id: Idd5d4567cfe25dd1e11a02995286a29606eac6e0

Former-commit-id: 30b4d18bae3d5261f10399a41a6682d80457a408