Set parent for WindowProcessController
If WindowProcessController was registered to display configuration
changes, then an override from display level can erase some of the
configuration fields. This can result in incomplete configuration.
Since process configuration is used as a "global" config in some
places, it should always be complete and contain all fields.

This CL uses ActivityRootContainer as parent for a
WindowProcessController, so any overrides should be applied on top
of global configuration and the resulting full config will always be
complete.

Bug: 131915789
Test: WindowProcessControllerTests#testConfigurationForSecondaryScreen
Change-Id: Id4d30ec9fc64dbdf3f31347240f6bb3cbd655fb2

Former-commit-id: 33ab86910e9090d10dd91d760d2dc53e1f3e99e6