Fixed an issue where the notification could be tiny
When updating the notificationWrapper we were reapplying
the translation from the row to keep the header hidden
during headsup. This was applied even on the public layout.
Because we didn't update all layouts though when the notification
stops being heads upped, the public layout never received the
update.
We're fixing this by not allowing the translation
on the header views and additionally also letting all content
views know of the state change.

Fixes: 116864754
Test: add HUN, update it, observe lockscreen (with redaction)
Change-Id: I1c9d1dab61081fcf2ff9a4dc22a741e565076294

Former-commit-id: 4b259a16af060ab3f4f4c09cfe18bf31008aa314