Don't clear source bounds when set to null during an animation.
There are several places where sysui will re-adjust the pip bounds.
When doing so, it accidentally clears the source bounds the app set when
entering pip. This causes the enter animation to look incorrect.

This change will not clear the source bounds if set to null during an
animation. They will get set the next time the pinned stack is resized.

Test: Youtube entering PIP crops correctly instead of scaling
Fixes: 132729493
Change-Id: Ifa247c2e2fff1b608a977762939570bfdf3f9f1b

Former-commit-id: 01f1aa4bb43591ce850b860ed2ef27b916833210