Not showing menu buttons for any headsup notification
Since swipe to remove notification behavior was rolled back to P which allows user to
remove a notification in both direction. We should not showing menu
buttons for any heads up notification either as in P.

Now it depends on the mIsUsingBidirectionalSwipe flag of NotificationMenuRow
to show menu buttons & the background.

Bug: 132668574
Test: send headsup notifications and swipe left/right
case and swipe left/right to check the background.

Change-Id: I5b96faed05d6184e02a8ede08f5e7c04ba5c456a

Former-commit-id: 3a950c371648ef20600a951cac11507fb7a1f190